requirements.txt
Flask==3.1.0
waitress==3.0.2
langchain==0.3.19
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.13.3
mistralai==1.5.0
faiss-cpu==1.9.0
numpy==1.26.4
sentence-transformers==3.4.1
googletrans==4.0.2
